requirements.txt
Django==2.2.5
djangorestframework==3.10.3
markdown==3.1.1
django-filter==2.2.0
djangorestframework-simplejwt==4.3.0
coverage==4.5.4
django-rest-swagger==2.2.0
